skim-back/requirements.txt
fastapi 
httpx 
python-dotenv 
uvicorn
psycopg2